Refactoring testAddingItemToCart And testLogoutUser
